singlevm: Modify singlevm experience to use new deployment tools
The new deployment tools are used to: a.) generate the SSNTP
certificates b.) install and manage the binaries via systemd

As part of this the cleanup scripts were also modified in order to
cleanup via systemd where necessary.

./setup.sh now runs a cleanup automatically based off the version saved
into the ~/local directory.

+++ testutil/singlevm/setup.sh
@@ -7,11 +7,12 @@ ciao_vlan_ip=198.51.100.1
 ciao_vlan_subnet=${ciao_vlan_ip}/24
 ciao_vlan_brdcast=198.51.100.255
 ciao_bin="$HOME/local"
-ciao_cert="$ciao_bin""/cert-Scheduler-""$ciao_host"".pem"
 keystone_key="$ciao_bin"/keystone_key.pem
 keystone_cert="$ciao_bin"/keystone_cert.pem
 workload_sshkey="$ciao_bin"/testkey
 ciao_pki_path=/etc/pki/ciao
+ciao_cert="$ciao_pki_path""/cert-Scheduler-""$ciao_host"".pem"
+
 export no_proxy=$no_proxy,$ciao_vlan_ip,$ciao_host
 
 ciao_email="ciao-devel@lists.clearlinux.org"
@@ -21,6 +22,8 @@ ciao_gobin="$GOPATH"/bin
 ciao_scripts="$GOPATH"/src/github.com/01org/ciao/testutil/singlevm
 ciao_env="$ciao_bin/demo.sh"
 ciao_dir=/var/lib/ciao
+ciao_data_dir=${ciao_dir}/data
+ciao_ctl_dir=${ciao_data_dir}/controller
 ciao_cnci_image="clear-8260-ciao-networking.img"
 ciao_cnci_url="https://download.clearlinux.org/demos/ciao"
 fedora_cloud_image="Fedora-Cloud-Base-24-1.2.x86_64.qcow2"
@@ -90,6 +93,9 @@ done
 
 echo "Subnet =" $ciao_vlan_subnet
 
+echo "Cleaning up"
+"$ciao_bin"/cleanup.sh 2> /dev/null 1>/dev/null
+
 # Copy the cleanup scripts
 cp "$ciao_scripts"/cleanup.sh "$ciao_bin"
 
@@ -144,40 +150,11 @@ set -o nounset
 
 echo "Generating workload ssh key $workload_sshkey"
 rm -f "$workload_sshkey" "$workload_sshkey".pub
-ssh-keygen -f "$workload_sshkey" -t rsa -N ''
+ssh-keygen -f "$workload_sshkey" -t rsa -N '' > /dev/null
 test_sshkey=$(< "$workload_sshkey".pub)
 chmod 600 "$workload_sshkey".pub
 #Note: Password is set to ciao
 test_passwd='$6$rounds=4096$w9I3hR4g/hu$AnYjaC2DfznbPSG3vxsgtgAS4mJwWBkcR74Y/KHNB5OsfAlA4gpU5j6CHWMOkkt9j.9d7OYJXJ4icXHzKXTAO.'
-
-echo "Generating configuration file $conf_file"
-(
-cat <<-EOF
-configure:
-  scheduler:
-    storage_uri: /etc/ciao/configuration.yaml
-  storage:
-    ceph_id: ciao
-  controller:
-    compute_ca: $keystone_cert
-    compute_cert: $keystone_key
-    identity_user: ${ciao_username}
-    identity_password: ${ciao_password}
-    cnci_vcpus: 4
-    cnci_mem: 128
-    cnci_disk: 128
-    admin_ssh_key: ${test_sshkey}
-    admin_password: ${test_passwd}
-  launcher:
-    compute_net: [${ciao_vlan_subnet}]
-    mgmt_net: [${ciao_vlan_subnet}]
-    disk_limit: false
-    mem_limit: false
-  identity_service:
-    type: keystone
-    url: https://${ciao_host}:${keystone_admin_port}
-EOF
-) > $conf_file
 
 sudo mkdir -p ${ciao_pki_path}
 if [ ! -d ${ciao_pki_path} ]
@@ -192,19 +169,13 @@ then
 	echo "FATAL ERROR: Unable to create /etc/ciao"
 	exit 1
 fi
-sudo install -m 0644 -t /etc/ciao $conf_file
 
-#Stop any running agents and CNCIs
-sudo killall ciao-scheduler
-sudo killall ciao-controller
-sudo killall ciao-launcher
-sudo killall qemu-system-x86_64
-sudo rm -rf ${ciao_dir}
-
-cd "$ciao_bin"
-
-#Cleanup any old artifacts
-rm -f "$ciao_bin"/*.pem
+sudo mkdir -p ${ciao_ctl_dir}/workloads
+if [ ! -d ${ciao_ctl_dir}/workloads ]
+then
+	echo "FATAL ERROR: Unable to create ${ciao_ctl_dir}/workloads}"
+	exit 1
+fi
 
 #Build ciao
 rm -f "$ciao_gobin"/ciao*
@@ -217,8 +188,6 @@ then
 	exit 1
 fi
 
-cd "$ciao_bin"
-
 #Check if the build was a success
 if [ ! -f "$ciao_gobin"/ciao-cli ]
 then
@@ -229,21 +198,14 @@ fi
 # Started early to minimise overall running time
 source $ciao_scripts/setup_keystone.sh
 
-#Generate Certificates
-"$GOPATH"/bin/ciao-cert -anchor -role scheduler -email="$ciao_email" \
-    -organization="$ciao_org" -host="$ciao_host" -ip="$ciao_vlan_ip" -verify
-
-"$GOPATH"/bin/ciao-cert -role cnciagent -anchor-cert "$ciao_cert" \
-    -email="$ciao_email" -organization="$ciao_org" -host="$ciao_host" \
-    -ip="$ciao_vlan_ip" -verify
-
-"$GOPATH"/bin/ciao-cert -role controller -anchor-cert "$ciao_cert" \
-    -email="$ciao_email" -organization="$ciao_org" -host="$ciao_host" \
-    -ip="$ciao_vlan_ip" -verify
-
-"$GOPATH"/bin/ciao-cert -role agent,netagent -anchor-cert "$ciao_cert" \
-    -email="$ciao_email" -organization="$ciao_org" -host="$ciao_host" \
-    -ip="$ciao_vlan_ip" -verify
+killall ciao-deploy-server
+"$ciao_gobin"/ciao-deploy-server -admin-password="$test_passwd" \
+    -admin-ssh-key="$test_sshkey" -ceph-id="ciao" \
+    -compute-net="$ciao_vlan_subnet" -https-ca-cert="$keystone_cert" \
+    -https-cert="$keystone_key" -keystone-password="$ciao_password" \
+    -keystone-url="https://$ciao_host:$keystone_admin_port" \
+    -keystone-user="$ciao_username" -mgmt-net="$ciao_vlan_subnet" \
+    -scheduler -server-ip="$ciao_vlan_ip" &
 
 source $ciao_scripts/setup_webui.sh
 
@@ -288,17 +250,12 @@ sudo ceph auth get-or-create client.ciao -o /etc/ceph/ceph.client.ciao.keyring m
 
 source "$ciao_scripts"/wait_for_keystone.sh
 
-#Copy the launch scripts
-cp "$ciao_scripts"/run_scheduler.sh "$ciao_bin"
-cp "$ciao_scripts"/run_controller.sh "$ciao_bin"
-cp "$ciao_scripts"/run_launcher.sh "$ciao_bin"
 cp "$ciao_scripts"/verify.sh "$ciao_bin"
 
-#Kick off the agents
-cd "$ciao_bin"
-"$ciao_bin"/run_scheduler.sh  &> /dev/null
-"$ciao_bin"/run_launcher.sh   &> /dev/null
-"$ciao_bin"/run_controller.sh &> /dev/null
+"$GOPATH"/bin/ciao-deploy-client -deploy-server="http://$ciao_host:9000" -force \
+    -role controller
+"$GOPATH"/bin/ciao-deploy-client -deploy-server="http://$ciao_host:9000" -force \
+    -role dual   
 
 # become admin in order to upload images and setup workloads
 export CIAO_USERNAME=$CIAO_ADMIN_USERNAME
